requirements.txt
psutil==5.8.0
Pillow==8.2.0
pyglet==1.5.16
opensimplex==0.3
clipboard==0.0.4
toml==0.10.2
requests==2.25.1
deprecation==2.1.0
simplejson==3.17.2
black==21.4b2
isort==5.8.0
Python-NBT==1.2.0